NVIDIA: SAUCE: arm64: RMI: Support for the VGIC in realms
The RMM provides emulation of a VGIC to the realm guest. With RMM v2.0
the registers are passed in the system registers so this works similar
to a normal guest, but kvm_arch_vcpu_put() need reordering to early out,
and realm guests don't support GICv2 even if the host does.

Signed-off-by: Steven Price <steven.price@arm.com>
[cherry-picked from cca-host/v13 1d43924f5992]
Signed-off-by: Ian May <ianm@nvidia.com>