NVIDIA: SAUCE: arm64: RMI: Runtime faulting of memory
At runtime if the realm guest accesses memory which hasn't yet been
mapped then KVM needs to either populate the region or fault the guest.

For memory in the lower (protected) region of IPA a fresh page is
provided to the RMM which will zero the contents. For memory in the
upper (shared) region of IPA, the memory from the memslot is mapped
into the realm VM non secure.

Signed-off-by: Steven Price <steven.price@arm.com>
[backported from cca-host/v13 2c68c568403f]
[context: adjusted for e07fc9e2da91 ("KVM: arm64: Fix page leak in
user_mem_abort() on atomic fault") which refactored error handling from
inline kvm_release_page_unused()+return to goto out_put_page. Kept the
goto pattern and added the realm protected address check before it.]
Signed-off-by: Ian May <ianm@nvidia.com>