NVIDIA: SAUCE: arm64: RMI: Activate realm on first VCPU run
Use kvm_arch_vcpu_run_pid_change() to check if this is the first time
the realm guest has run. If this is the first run then activate the
realm.

Before the realm can be activated it must first be created, this is a
stub in this patch and will be filled in by a later patch.

Signed-off-by: Steven Price <steven.price@arm.com>
[cherry-picked from cca-host/v13 73b6917c99fb]
Signed-off-by: Ian May <ianm@nvidia.com>