NVIDIA: SAUCE: arm64: RMI: Enable PMU support with a realm guest
Use the PMU registers from the RmiRecExit structure to identify when an
overflow interrupt is due and inject it into the guest. Also hook up the
configuration option for enabling the PMU within the guest.

The number of PMU counters is configured by the VMM by writing to PMCR.N.

Signed-off-by: Steven Price <steven.price@arm.com>
[cherry-picked from cca-host/v13 1bdad244f42f]
Signed-off-by: Ian May <ianm@nvidia.com>