arm_mpam: Add force-disable debugfs trigger
MPAM has an error interrupt that can be triggered by an MSC when
corrupt or out of range values are seen. The hardware only needs to
raise an error interrupt if the error was detected, it is also
permissible for the hardware to just use the corrupt or our of range
value. All the reasons to raise an error indicate a software bug.

When the error interrupt is triggered, the MPAM driver
attempts to reset all the CPUs back to PARTID-0 and reset PARTID-0
to be unrestricted. This is done to ensure important tasks aren't
accidentally given the performance of unimportant tasks.

This teardown path in the driver is hard to trigger. Add a debugfs
file to poke this manually. It is expected you have to reboot to
make MPAM work again after this.

Signed-off-by: James Morse <james.morse@arm.com>
(cherry picked from commit 8c96f858b25aa42694c5db56a2afe255ed8262dd https://web.git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.14-rc1)
Signed-off-by: Koba Ko <kobak@nvidia.com>
Reviewed-by: Fenghua Yu <fenghuay@nvidia.com>
Reviewed-by: Tushar Dave <tdave@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>