DOWNSTREAM: arm64: rme: Enable earlycon for cca guest
The earlycon feature is not officially supported in the upstream
kernel. However, the kernel crashes if the user passes the
'earlycon' boot argument. To avoid this, the user must either
provide the 'earlycon' argument with a shared MMIO address or
refrain from enabling earlycon, which creates inconvenience for
internal development and testing. To address this, a workaround
is added to the kernel to prevent crashes.

Change-Id: I422b8c17f647fa71ffc7febf4e57edf5b312c225
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>