NVIDIA: SAUCE: WAR: acpi: arm64: set ACPI_IORT_MF_CANWBS
This flag is required to enable SMMU's nested translation feature, so as
to run UVM test case in the guest OS. Because the IORT firmwares are not
updated yet to enable the flag. Work it around by setting in the kernel.
And there should not be any side effect.

The IORT firmware should set this flag for PCI buses, to notify SMMU that
all devices under the PCI bus are cache coherent.

Change-Id: Ie560d7da5d7ae67be04499f8408188149a665d32
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
(cherry picked from commit 2b644f66bdcfa91e6cbcbefbe835005417ae4979 dev-tb500-nvidia-pset-linux-6.16)
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Nirmoy Das <nirmoyd@nvidia.com>
Acked-by: Carol L. Soto <csoto@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>