NVIDIA: SAUCE: PCI: Add pci_bus_isolated()
Prepare to move the calculation of the bus P2P isolation out of the iommu
code and into the PCI core. This allows using the faster list iteration
under the pci_bus_sem, and the code has a kinship with the logic in
pci_for_each_dma_alias().

Isolation is the concept that drives the iommu_groups for the purposes of
VFIO. Stated simply, if device A can send traffic to device B then they
must be in the same group. An isolated device is one where its
transactions are only delivered to the IOMMU.

There are three areas of isolation that impact how transactions can reach
something other than the IOMMU:

 - Bus Isolation. Upstream directed transactions mastered to a bus can go
   to a downstream slave. In classical PCI buses are not isolating
   because the multi-drop electrical topology allows any device to act as
   slave. Physical PCIe links are isolating because they are point to
   point.

 - Switch Iternal Bus Isolation. This is a special case of Bus Isolation
   for PCIE where ACS on the Downstream Ports of a switch controls how
   transactions flow on the internal bus to prevent other Downstream Ports
   from claming them.

 - Bridge (i.e. Switch DSP/USP, Root Port) Loopback Isolation. If the
   bridge has BARs then transactions from the downstream bus can land on the
   BARs of bridge without reaching a bus.

 - MFD Internal Loopback Isolation. Transactions from one MFD function can
   be internally routed to another function without reaching a bus.

pci_bus_isolated() is concerned with the first three. Later patches in
this series deal with the last one.

A bus is considered isolating if upstream directed transactions mastered
on the bus always continue to travel upstream. If the bus permits upstream
directed transactions to travel back downstream then it is a non-isolating
bus. Some examples:

 - The point to point "bus" on a physical PCIe link is isolated if the
   bridge/root device has something preventing self-access to its own
   MMIO.

 - The downstream bus from a Root Port is usually isolated from the rest
   of the Root Complex.

 - The internal bus of a PCIe switch can be isolated if all its Downstream
   Ports have correct ACS flags

pci_bus_isolated() implements these rules and returns an enum indicating
the level of isolation the bus has, with five possibilies:

 PCIE_ISOLATED: Transactions mastered on the bus headed in the upstream
      direction always continue upstream

 PCIE_SWITCH_DSP_NON_ISOLATED: The bus is the internal bus of a PCIE
      switch, transactions cannot reach any BARs of the USP bus but the DSPs
      can claim upstream directed transactions.

 PCIE_NON_ISOLATED: The bus is the internal bus of a PCIE switch and
     upstream directed transactions can reach the BARs of the USP or be
     claimed by the DSPs.

 PCI_BUS_NON_ISOLATED: The bus is PCI/PCI-X but the bridge is PCIe, has no
     aliases and the bridge's BARs do not claim upstream directed
     transactions.

 PCI_BRIDGE_NON_ISOLATED: The bus is PCI/PCI-X and has no isolation, and
     the bridges's BARs can claim upstream directed transactions.

The calculation is done per-bus, so it is possible for a transaction from
a PCI device to travel through different bus isolation types on its way
upstream. PCIE_SWITCH_DSP_NON_ISOLATED/PCI_BUS_NON_ISOLATED and
PCIE_NON_ISOLATED/PCI_BRIDGE_NON_ISOLATED are the same for the purposes of
creating iommu groups. The distinction between PCIe and PCI allows for
easier understanding and debugging as to why the groups are chosen.

For the iommu groups if all buses on the upstream path are PCIE_ISOLATED
then the end device has a chance to have a single-device iommu_group. Once
any non-isolated bus segment is found that bus segment will have an
iommu_group that captures all downstream devices, and sometimes the
upstream bridge.

pci_bus_isolated() is principally about isolation, but there is an
overlap with grouping requirements for legacy PCI aliasing. For purely
legacy PCI environments pci_bus_isolated() returns
PCI_BRIDGE_NON_ISOLATED for everything and all devices within a hierarchy
are in one group. No need to worry about bridge aliasing.

Reviewed-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
(cherry picked from commit 516f3934ddf7189eb3e137e05ef645a5b9b8e83b https://github.com/jgunthorpe/linux/commits/pcie_switch_groups)
Signed-off-by: Nikolay Aleksandrov <nikolay@nvidia.com>
Acked-by: Nirmoy Das <nirmoyd@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>